requirements.txt
feedparser>=6.0.0
numpy>=1.20.0
openai>=1.0.0
pandas>=1.3.0
pytz==2023.3
requests>=2.25.1
yfinance>=0.2.12
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv>=1.0.0
fredapi>=0.5.0
python-dateutil>=2.8.2
colorama>=0.4.4
matplotlib>=3.5.0
httpx==0.27.2 